First commit was messed up, second one should do it.
I created a class Bacteria in bacteria.py , which is able to generate bacteria with parameters position, length, height, angle, growth_rate, velocity, and index. Also some functions are implemented which move the bacteria in direction of velocity and add some random movement and random rotation. The bacteria is able to split and evolve into two new bacteria with different parameters. The print_info function with prints the current parameters of a bacteria and the evolution function, which combines growth, movement and splitting.

The test_run.py script loads the bacteria class and tests its function.